admin section controller, routes, views refactor

+++ lib/tasks/ls.rake
@@ -3,8 +3,8 @@ namespace :ls do
     input = ''
     STDOUT.puts "Drop, recreate, and seed database? y[es] or n[o]"
     input = STDIN.gets.chomp
-    if input == "y"
-      Rake::Task["db:drop"].execute
+    if input == "y" && Rails.env == 'development'
+      Rake::Task["db:drop"].execute 
       Rake::Task["db:create"].execute
       Rake::Task["db:schema:load"].execute
       Rake::Task["db:seed"].execute
